Use unsigned in preference to uint64_t for element counts and field widths
git-svn-id: http://parabix.costar.sfu.ca/svn/icGREP/icgrep-devel@5231 a82c6ad6-7e1c-4be0-8e68-26c050b43eaf

+++ icgrep/IDISA/idisa_builder.cpp
@@ -574,7 +574,7 @@ StoreInst * IDISA_Builder::CreateAtomicStoreRelease(Value * val, Value * ptr) {
     return inst;
 }
 
-Type * IDISA_Builder::getStreamTy(const uint64_t FieldWidth) {
+Type * IDISA_Builder::getStreamTy(const unsigned FieldWidth) {
     const auto f = mStreamTypes.find(FieldWidth);
     if (LLVM_LIKELY(f != mStreamTypes.end())) {
         return f->second;
