+++ icgrep/pablo/codegenstate.h
@@ -257,11 +257,11 @@ public:
 
     While * createWhile(PabloAST * condition, PabloBlock * body);
 
-    Type * getStreamTy(const uint64_t FieldWidth = 1) {
+    Type * getStreamTy(const unsigned FieldWidth = 1) {
         return mParent->getStreamTy(FieldWidth);
     }
     
-    Type * getStreamSetTy(const uint64_t NumElements = 1, const uint64_t FieldWidth = 1) {
+    Type * getStreamSetTy(const unsigned NumElements = 1, const unsigned FieldWidth = 1) {
         return mParent->getStreamSetTy(NumElements, FieldWidth);
     }
